🩹 hot-patch bma4xx driver at startup to handle 12-bit acceleration sensor values (#27)
The in-tree bma4xx driver does not handle the 12bit sensor values from
the bma423 correctly. This patches the decode method at startup to
convert the 12-bit values to 16-bit values. It is technically invoking
undefined behavior (via a write to a const struct), so this should be
watched with caution.

The real fix is to modify the in-tree bma4xx driver to handle the 12 bit
values when the bma423 is used.